pallet_transporter

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

CancelledTransfers

Storage to track cancelled transfers between different chains.

ChainTransfers

A temporary storage that tracks total transfers from this chain. Clears on on_initialize for every block.

DomainBalances

Domain balances.

OutgoingTransfers

All the outgoing transfers on this execution environment.

UnconfirmedTransfers

Storage to track unconfirmed transfers between different chains.