pallet_transporter
Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type SelfChainId: Get<ChainId>;
    type SelfEndpointId: Get<EndpointId>;
    type Currency: Currency<Self::AccountId>;
    type Sender: Sender<Self::AccountId>;
    type AccountIdConverter: TryConvertBack<Self::AccountId, MultiAccountId>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Event type for this pallet.

type SelfChainId: Get<ChainId>

Gets the chain_id of the current execution environment.

type SelfEndpointId: Get<EndpointId>

Gets the endpoint_id of this pallet in a given execution environment.

type Currency: Currency<Self::AccountId>

Currency used by this pallet.

type Sender: Sender<Self::AccountId>

Sender used to transfer funds.

type AccountIdConverter: TryConvertBack<Self::AccountId, MultiAccountId>

MultiAccountID <> T::AccountId converter.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Object Safety

This trait is not object safe.