Crate pallet_transporter

Pallet transporter used to move funds between domains.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

weights

Autogenerated weights for pallet_transporter

Structs

EndpointHandler

Endpoint handler implementation for pallet transporter.

Location

Location that either sends or receives transfers between domains.

Pallet

Pallet transporter to move funds between domains.

Transfer

Transfer of funds from one domain to another.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

Errors emitted by pallet-transporter.

Event

Events emitted by pallet-transporter.

Traits

Config

Configuration trait of this pallet.

Type Definitions

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.